Rework folder detection
Utilize the Falcon BMS shared memory to get the Falcon BMS and briefing
folders. This is more robust and it removes the need to read the
registry.
As a side effect, I have removed the OS specific parts. I haven't tested
Falcon BMS on Linux, but I assume it should work now as well.